// src/entities/user/applications/model/application.enums.ts
export enum CraneTypes {
  Bridge = 'Bridge', // "Ko'prikli",
  Crawler = 'Crawler', // "O'zi sudralib yuruvchi",
  Funicular = 'Funicular', // "Funikulyar",
  Gantry = 'Gantry', // "Kozlovoy",
  Manipulator = 'Manipulator', // "Manipulyator",
  PipeLayer = 'PipeLayer', // "Truboukladchik";
  TireTrimmed = 'TireTrimmed', // "Pnevmoxodli",
  Tower = 'Tower', // "Minorali",
  Truck = 'Truck', //"Avtokran",
}

export enum ContainerTypes {
  Steam = 'Steam', // "Bug'li",
  AccumulatorCylinder = 'AccumulatorCylinder', // "Akkumlyator balloni",
  MobileVessel = 'MobileVessel', // "Harakatlanuvchi idish",
  AirCollector = 'AirCollector', // "Havo yig'gich",
  Receiver = 'Receiver', // "Qabul qiluvchi",
  Separator = 'Separator', // "Separator";
}